Raylet actor starts on dashboard

+++ dashboard/modules/logical_view/logical_view_head.py
@@ -9,18 +9,19 @@ from ray.core.generated import core_worker_pb2
 from ray.core.generated import core_worker_pb2_grpc
 import subprocess
 import random
-from typing import Dict
+from typing import Dict, Tuple
 from grpc.experimental import aio as aiogrpc
 
 logger = logging.getLogger(__name__)
 routes = dashboard_utils.ClassMethodRouteTable
 
+StreamlitConnection = Tuple[int, subprocess.Popen]
 
 class LogicalViewHead(dashboard_utils.DashboardHeadModule):
     def __init__(self, dashboard_head):
         super().__init__(dashboard_head)
         initial_actors = DataSource.actors
-        self.streamlit_servers: Dict[str, subprocess.Popen] = {}
+        self.streamlit_servers: Dict[str, StreamlitConnection] = {}
 
         for actor_id, actor in initial_actors.items():
             logger.info("Starting an initial streamlit server")
@@ -43,25 +44,37 @@ class LogicalViewHead(dashboard_utils.DashboardHeadModule):
         # TODO(hackathon) Replace with link to script
         logger.info("Starting a streamlit server")
         port_number = random.randint(49152, 65535)
-        streamlit_command = [
-            "streamlit", "run", "--server.port", f"{port_number}",
-            "/Users/maxfitton/Development/ray/ray_streamlit_actor.py", "--",
-            "--actor-name=ActorA"
-        ]
-        proc = subprocess.Popen(streamlit_command)
-        # logger.info(f"PROCESS={proc.__dict__}")
-        # logger.info(f"PROCESS OUTPUT LOG = {proc.stdout}")
-        # logger.info(f"PROCESS OUTPUT ERR = {proc.stderr}")
-        self.streamlit_servers[actor_id] = (port_number, proc)
+        actor_name = actor["name"]
+        script_path = actor["streamlitScriptPath"]
+        if script_path:
+            streamlit_command = [
+                "streamlit", "run", "--server.port", f"{port_number}",
+                actor["streamlitScriptPath"], "--",
+                f"--actor-name={actor_name}"
+            ]
+            proc = subprocess.Popen(streamlit_command)
+            # logger.info(f"PROCESS={proc.__dict__}")
+            # logger.info(f"PROCESS OUTPUT LOG = {proc.stdout}")
+            # logger.info(f"PROCESS OUTPUT ERR = {proc.stderr}")
+            self.streamlit_servers[actor_id] = (port_number, proc)
 
     def stop_streamlit_server(self, actor_id, actor):
         logger.info("Killing a streamlit server")
-        streamlit_process = self.streamlit_servers[actor_id]
-        streamlit_process.kill()
+        script_path = actor["streamlitScriptPath"]
+        if script_path:
+            streamlit_process = self.streamlit_servers[actor_id]
+            streamlit_process.kill()
 
     @routes.get("/logical/actor_groups")
     async def get_actor_groups(self, req) -> aiohttp.web.Response:
         actors = await DataOrganizer.get_all_actors()
+        for actor in actors.values():
+            streamlit_script_path = actor["streamlitScriptPath"]
+            if streamlit_script_path != "":
+                conn = self.streamlit_servers.get(actor["actorId"])
+                if conn:
+                    port, _ = conn
+                    actor["streamlitPort"] = port
         actor_creation_tasks = await DataOrganizer.get_actor_creation_tasks()
         # actor_creation_tasks have some common interface with actors,
         # and they get processed and shown in tandem in the logical view
